backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
prometheus-client==0.20.0
redis==5.0.4
psycopg2-binary==2.9.9
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
